+++ sports/soccer/annotators.py
@@ -4,7 +4,7 @@ import cv2
 import supervision as sv
 import numpy as np
 
-from sports.configs.soccer import SoccerPitchConfiguration
+from sports.soccer.config import SoccerPitchConfiguration
 
 
 def draw_pitch(
