+++ sports/soccer/config.py
